Correções nas rotas

+++ src/main/java/com/api/crudspring/controller/CursoController.java
@@ -5,6 +5,7 @@ import java.util.List;
 import com.api.crudspring.model.Course;
 import com.api.crudspring.service.CourseService;
 
+import org.apache.catalina.connector.Response;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -22,7 +23,7 @@ import lombok.AllArgsConstructor;
 // O @RequestMapping indica a URL do controller
 @RestController
 @AllArgsConstructor
-@RequestMapping("/api/cursos")
+@RequestMapping("/api")
 public class CursoController {
 
     private final CourseService courseService;
@@ -30,31 +31,42 @@ public class CursoController {
     // As anotações de mapping indicam o endereço para chamada do método
     @GetMapping("")
     public String index() {
-        return "Exemplo de API REST - Java Spring";
+        return "<h1>Exemplo de API REST - Java Spring</h1>Java Spring, JPA Hibernate e PostgreSQL<span></span>";
     }
 
-    @GetMapping("/list")
-    public List<Course> listCourse() {
-        return this.courseService.findAllCourses();
-    }
-
-    @GetMapping("/{id}")
-    public Course findCourseById(@PathVariable(value = "id") long id) {
-        return this.courseService.findCourseById(id);
-    }
-
-    @PostMapping("/save")
+    // Salvar curso
+    @PostMapping("cursos")
     public Course saveCourse(@RequestBody Course course) {
         return this.courseService.saveCourse(course);
     }
 
-    @PutMapping("/save")
-    public Course updateCourse(@RequestBody Course course) {
+    // Listar cursos
+    @GetMapping("cursos")
+    public List<Course> listCourse() {
+        return this.courseService.findAllCourses();
+    }
+
+    // Atualizar curso
+    @PutMapping("cursos/{id}")
+    public Course updateCourse(@PathVariable(value = "id") long id, @RequestBody Course course) {
+        if (id != course.getId()) {
+            return null;
+        }
         return this.courseService.saveCourse(course);
     }
 
-    @DeleteMapping("/delete")
-    public void deleteCourse(@RequestBody Course course) {
+    // Recuperar curso por id
+    @GetMapping("cursos/{id}")
+    public Course findCourseById(@PathVariable(value = "id") long id) {
+        return this.courseService.findCourseById(id);
+    }
+
+    // Remover curso
+    @DeleteMapping("cursos/{id}")
+    public void deleteCourse(@PathVariable(value = "id") long id, @RequestBody Course course) {
+        if (id != course.getId()) {
+            return;
+        }
         this.courseService.deleteCourse(course);
     }
